% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_full.R
\name{linear_regression_full}
\alias{linear_regression_full}
\title{Run a linear regression analysis of two chosen variables in a dataset, check normality, then print resulting graph to screen.}
\usage{
linear_regression_full(variable_1, variable_2)
}
\arguments{
\item{variable_1:}{response variable (y-axis/dependent)}

\item{varible_2:}{predictor variable (x-axis/independent)}
}
\value{
qqline graph printed to screen
}
\description{
Run a linear regression analysis of two chosen variables in a dataset, check normality, then print resulting graph to screen.
}
