% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_anova.R
\name{simple_anova}
\alias{simple_anova}
\title{Run an ANOVA for chosen variables in a dataset, then print summary to screen.}
\usage{
simple_anova(variable_1, variable_2)
}
\arguments{
\item{variable_1:}{response variable (y-axis/dependent)}

\item{varible_2:}{categorical predictor variable (x-axis/independent)}
}
\value{
table of summary statistics printed to screen
}
\description{
Run an ANOVA for chosen variables in a dataset, then print summary to screen.
}
